+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-actions.ts
@@ -11,6 +11,7 @@ export const canadaHierarchyIndex = [0, 1, 0, 4];
 export const togoHierarchy = ['Africa', 'Togo', 'Lome'];
 export const hierarchicalField = 'geographicalhierarchy';
 export const defaultNumberOfValues = 5;
+export const categoryFacetLabel = 'Atlas';
 
 export interface CategoryFacetSetupOptions {
   field: string;