+++ packages/atomic/cypress/integration/result-list/result-components/result-text.cypress.ts
@@ -1,4 +1,4 @@
-import {HighlightKeyword} from '@coveo/headless';
+import {HighlightKeyword, Result} from '@coveo/headless';
 import {
   generateComponentHTML,
   TagProps,
@@ -109,7 +109,9 @@ describe('Result Text Component', () => {
           response.results.forEach((result) => {
             result.raw[field] = rawValue;
             if (highlightsAvailable) {
-              (result as any)[`${field}Highlights`] = [highlight];
+              (result as Result & Record<string, HighlightKeyword[]>)[
+                `${field}Highlights`
+              ] = [highlight];
             }
           })
         )
@@ -126,6 +128,8 @@ describe('Result Text Component', () => {
         ResultTextSelectors.firstInResult().should('have.text', rawValue);
         ResultTextSelectors.highlight().should('have.text', highlightedValue);
       });
+
+      CommonAssertions.assertAccessibility(ResultTextSelectors.firstInResult);
     });
 
     describe('when the "shouldHighlight" prop is false', () => {