+++ packages/atomic/src/components/atomic-search-box/atomic-search-box.tsx
@@ -22,6 +22,11 @@ import {
   SearchBoxSuggestionsBindings,
   SearchBoxSuggestionsEvent,
 } from '../search-box-suggestions/suggestions-common';
+import {
+  AriaLiveRegion,
+  FocusTarget,
+  FocusTargetController,
+} from '../../utils/accessibility-utils';
 
 /**
  * The `atomic-search-box` component creates a search box with built-in support for suggestions.
@@ -66,6 +71,12 @@ export class AtomicSearchBox {
    *  - atomic-search-box-recent-queries
    */
   @Prop() public numberOfQueries = 8;
+
+  @AriaLiveRegion('search-box')
+  protected ariaMessage!: string;
+
+  @FocusTarget()
+  private inputFocus!: FocusTargetController;
 
   public initialize() {
     this.id = randomID('atomic-search-box-');
@@ -112,7 +123,7 @@ export class AtomicSearchBox {
       id: this.id,
       searchBoxController: this.searchBox,
       numberOfQueries: this.numberOfQueries,
-      inputRef: this.inputRef,
+      clearSuggestions: () => this.clearSuggestions(),
       triggerSuggestions: () => this.triggerSuggestions(),
       getSuggestions: () => this.suggestions,
     };
@@ -147,11 +158,11 @@ export class AtomicSearchBox {
   }
 
   private get firstValue() {
-    return this.listRef.firstElementChild!;
+    return this.listRef.firstElementChild;
   }
 
   private get lastValue() {
-    return this.listRef.lastElementChild!;
+    return this.listRef.lastElementChild;
   }
 
   private get nextOrFirstValue() {
@@ -179,7 +190,7 @@ export class AtomicSearchBox {
   }
 
   private focusNextValue() {
-    if (!this.hasSuggestions) {
+    if (!this.hasSuggestions || !this.nextOrFirstValue) {
       return;
     }
 
@@ -190,7 +201,7 @@ export class AtomicSearchBox {
   }
 
   private focusPreviousValue() {
-    if (!this.hasSuggestions) {
+    if (!this.hasSuggestions || !this.previousOrLastValue) {
       return;
     }
 
@@ -198,6 +209,16 @@ export class AtomicSearchBox {
     !isNullOrUndefined(query) && this.updateQuery(query);
     this.updateActiveDescendant(this.previousOrLastValue.id);
     this.scrollActiveDescendantIntoView();
+  }
+
+  private updateAriaMessage() {
+    this.ariaMessage = this.suggestionElements.length
+      ? this.bindings.i18n.t('query-suggestions-available', {
+          count: this.suggestionElements.filter(
+            (element) => element.query !== undefined
+          ).length,
+        })
+      : this.bindings.i18n.t('query-suggestions-unavailable');
   }
 
   private async triggerSuggestions() {
@@ -213,9 +234,11 @@ export class AtomicSearchBox {
       this.numberOfQueries +
       suggestionElements.filter((sug) => sug.query === undefined).length;
     this.suggestionElements = suggestionElements.slice(0, max);
+    this.updateAriaMessage();
   }
 
   private onInput(value: string) {
+    this.isExpanded = true;
     this.searchBox.updateText(value);
     this.updateActiveDescendant();
     this.triggerSuggestions();
@@ -226,7 +249,7 @@ export class AtomicSearchBox {
     this.triggerSuggestions();
   }
 
-  private onBlur() {
+  private clearSuggestions() {
     this.isExpanded = false;
     this.updateActiveDescendant();
     this.clearSuggestionElements();
@@ -241,7 +264,7 @@ export class AtomicSearchBox {
 
     this.searchBox.submit();
     this.updateActiveDescendant();
-    this.inputRef.blur();
+    this.clearSuggestions();
   }
 
   private updateQuery(query: string) {
@@ -259,7 +282,7 @@ export class AtomicSearchBox {
         this.onSubmit();
         break;
       case 'Escape':
-        this.onBlur();
+        this.clearSuggestions();
         break;
       case 'ArrowDown':
         e.preventDefault();
@@ -276,7 +299,10 @@ export class AtomicSearchBox {
     return (
       <input
         part="input"
-        ref={(el) => (this.inputRef = el as HTMLInputElement)}
+        ref={(el) => {
+          this.inputRef = el as HTMLInputElement;
+          this.inputFocus.setTarget(el);
+        }}
         role="combobox"
         aria-autocomplete="both"
         aria-haspopup="true"
@@ -292,7 +318,7 @@ export class AtomicSearchBox {
         class="h-full outline-none bg-transparent flex-grow px-4 py-3.5 text-neutral-dark placeholder-neutral-dark text-lg"
         value={this.searchBoxState.value}
         onFocus={() => this.onFocus()}
-        onBlur={() => this.onBlur()}
+        onBlur={() => this.clearSuggestions()}
         onInput={(e) => this.onInput((e.target as HTMLInputElement).value)}
         onKeyDown={(e) => this.onKeyDown(e)}
       />
@@ -335,6 +361,7 @@ export class AtomicSearchBox {
 
   private clearSuggestionElements() {
     this.suggestionElements = [];
+    this.ariaMessage = '';
   }
 
   private renderSuggestion(
@@ -391,6 +418,7 @@ export class AtomicSearchBox {
         part="submit-button"
         ariaLabel={this.bindings.i18n.t('search')}
         onClick={() => {
+          this.inputFocus.focusAfterSearch();
           this.searchBox.submit();
           this.clearSuggestionElements();
         }}