+++ packages/atomic/src/components/facets/facet-show-more-less/facet-show-more-less.tsx
@@ -11,6 +11,8 @@ interface FacetShowMoreProps {
   canShowMoreValues: boolean;
   onShowMore(): void;
   onShowLess(): void;
+  showMoreRef?: (element?: HTMLButtonElement) => void;
+  showLessRef?: (element?: HTMLButtonElement) => void;
 }
 
 export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
@@ -40,6 +42,7 @@ export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
         class={`${btnClasses} ${props.canShowLessValues ? '' : 'hidden'}`}
         aria-label={showLessFacetValues}
         onClick={() => props.onShowLess()}
+        ref={props.showLessRef}
       >
         <atomic-icon
           part="show-more-less-icon"
@@ -54,6 +57,7 @@ export const FacetShowMoreLess: FunctionalComponent<FacetShowMoreProps> = (
         class={`${btnClasses} ${props.canShowMoreValues ? '' : 'hidden'}`}
         aria-label={showMoreFacetValues}
         onClick={() => props.onShowMore()}
+        ref={props.showMoreRef}
       >
         <atomic-icon
           part="show-more-less-icon"